- Added standard library license header - Moved BasicMergeSubscription into operator package - Made ParallelOrderedMerge package private and final - Moved source parameter in ParallelTransformers - Made test java6 compatible
